do the same with disconnect client
