+++ gtk-module/src/main.rs
@@ -1,9 +1,9 @@
 use cli_module::{create_player, error_exit, spawn_clean_up_mut};
 #[cfg(any(windows, target_os = "linux", target_os = "macos"))]
 use controls_module::StatusReceiver;
-use disconnect_module::DisconnectClientConfig;
+use disconnect_module::{DisconnectClientConfig, spawn_disconnect};
 #[cfg(target_os = "linux")]
-use mpris_module::launch_mpris;
+use mpris_module::spawn_mpris;
 use std::sync::Arc;
 use tokio::sync::{broadcast, mpsc, watch};
 
@@ -55,7 +55,7 @@ pub async fn run() -> AppResult<()> {
     .await?;
 
     #[cfg(target_os = "linux")]
-    launch_mpris(&player, &exit_sender, "io.github.sofusa.qobine".to_string());
+    spawn_mpris(&player, &exit_sender, "io.github.sofusa.qobine".to_string());
 
     #[cfg(any(windows, target_os = "linux", target_os = "macos"))]
     {
@@ -89,44 +89,13 @@ pub async fn run() -> AppResult<()> {
     let (active_device_tx, active_device_rx) = watch::channel(Default::default());
     let (set_active_device_tx, set_active_device_rx) = mpsc::unbounded_channel();
 
-    let position_receiver = player.position();
-    let tracklist_receiver = player.tracklist();
-    let volume_receiver = player.volume();
-    let status_receiver = player.status();
-    let controls = player.controls();
-    let active_sender = player.active_sender();
-    let auto_play_receiver = player.auto_play();
-
-    let tracklist_sender = player.tracklist_sender();
-    let position_sender = player.position_sender();
-    let status_sender = player.status_sender();
-    let volume_sender = player.volume_sender();
-    let auto_play_sender = player.auto_play_sender();
-
-    tokio::spawn(async move {
-        if let Err(e) = disconnect_module::init(
-            config_rx,
-            controls,
-            tracklist_sender,
-            position_sender,
-            volume_sender,
-            auto_play_sender,
-            status_sender,
-            active_sender,
-            available_devices_tx,
-            active_device_tx,
-            position_receiver,
-            tracklist_receiver,
-            status_receiver,
-            volume_receiver,
-            auto_play_receiver,
-            set_active_device_rx,
-        )
-        .await
-        {
-            error_exit(e);
-        }
-    });
+    spawn_disconnect(
+        &player,
+        config_rx,
+        available_devices_tx,
+        active_device_tx,
+        set_active_device_rx,
+    );
 
     let client = client.clone();
