+++ tui-module/src/main.rs
@@ -2,10 +2,10 @@ use cli_module::{
     ConnectArgs, SharedArgs, SharedCommands, create_player, default_audio_cache,
     default_audio_quality, error_exit, get_client, handle_shared_commands, spawn_clean_up_mut,
 };
-use disconnect_module::DisconnectClientConfig;
+use disconnect_module::{DisconnectClientConfig, spawn_disconnect};
 use futures::executor::block_on;
 #[cfg(target_os = "linux")]
-use mpris_module::launch_mpris;
+use mpris_module::spawn_mpris;
 use std::sync::Arc;
 use tokio::sync::{broadcast, mpsc, watch};
 
@@ -79,7 +79,7 @@ pub async fn run() -> AppResult<()> {
     .await?;
 
     #[cfg(target_os = "linux")]
-    launch_mpris(&player, &exit_sender, "qobuz-player".to_string());
+    spawn_mpris(&player, &exit_sender, "qobuz-player".to_string());
 
     #[cfg(any(windows, target_os = "linux", target_os = "macos"))]
     {
@@ -148,46 +148,13 @@ pub async fn run() -> AppResult<()> {
     let (active_device_tx, active_device_rx) = watch::channel(Default::default());
     let (set_active_device_tx, set_active_device_rx) = mpsc::unbounded_channel();
 
-    {
-        let position_receiver = player.position();
-        let tracklist_receiver = player.tracklist();
-        let volume_receiver = player.volume();
-        let status_receiver = player.status();
-        let controls = player.controls();
-        let active_sender = player.active_sender();
-        let auto_play_receiver = player.auto_play();
-
-        let tracklist_sender = player.tracklist_sender();
-        let position_sender = player.position_sender();
-        let status_sender = player.status_sender();
-        let volume_sender = player.volume_sender();
-        let auto_play_sender = player.auto_play_sender();
-
-        tokio::spawn(async move {
-            if let Err(e) = disconnect_module::init(
-                config_rx,
-                controls,
-                tracklist_sender,
-                position_sender,
-                volume_sender,
-                auto_play_sender,
-                status_sender,
-                active_sender,
-                available_devices_tx,
-                active_device_tx,
-                position_receiver,
-                tracklist_receiver,
-                status_receiver,
-                volume_receiver,
-                auto_play_receiver,
-                set_active_device_rx,
-            )
-            .await
-            {
-                error_exit(e);
-            }
-        });
-    }
+    spawn_disconnect(
+        &player,
+        config_rx,
+        available_devices_tx,
+        active_device_tx,
+        set_active_device_rx,
+    );
 
     tokio::spawn(async move {
         if let Err(e) = tui_module::init(
